Most expensive table added, wired product links up to summary page

+++ SoldOutWeb/Models/SearchSummary.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 
 namespace SoldOutWeb.Models
 {
@@ -7,7 +6,7 @@ namespace SoldOutWeb.Models
     {
         public string Name { get; set; }
         public string Description { get; set; }
-        public int SearchID { get; set; }
+        public long SearchID { get; set; }
         public DateTime LastRun { get; set; }
         public int TotalResults { get; set; }
         public string Link { get; set; }
